requirements.txt
# Tesseract-OCR needs to be in path
# apt install tesseract-ocr
pytesseract
pillow
Flask
weasyprint
requests
beautifulsoup4